increase accuracy for spectrograms and annotations

+++ bats_ai/core/models/nabat/nabat_spectrogram.py
@@ -18,11 +18,11 @@ logger = logging.getLogger(__name__)
 class NABatSpectrogram(TimeStampedModel, models.Model):
     nabat_recording = models.ForeignKey(NABatRecording, on_delete=models.CASCADE)
     images = GenericRelation(SpectrogramImage)
-    width = models.IntegerField()  # pixels
-    height = models.IntegerField()  # pixels
-    duration = models.IntegerField()  # milliseconds
-    frequency_min = models.IntegerField()  # hz
-    frequency_max = models.IntegerField()  # hz
+    width = models.FloatField()  # pixels
+    height = models.FloatField()  # pixels
+    duration = models.FloatField()  # milliseconds
+    frequency_min = models.FloatField()  # hz
+    frequency_max = models.FloatField()  # hz
 
     @property
     def image_url_list(self):
